Why am I doing this.
