+++ moq-karp/Cargo.toml
@@ -15,6 +15,7 @@ categories = ["multimedia", "network-programming", "web-programming"]
 
 [dependencies]
 moq-transfork = { path = "../moq-transfork", version = "0.9" }
+moq-transfork-proto = { path = "../moq-transfork-proto", version = "0.1" }
 moq-async = { path = "../moq-async", version = "0.1" }
 
 url = "2"
