+++ moq-transfork-proto/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "moq-transfork-proto"
+description = "Media over QUIC"
+authors = ["Luke Curley"]
+repository = "https://github.com/kixelated/moq-rs"
+license = "MIT OR Apache-2.0"
+
+version = "0.1.0"
+edition = "2021"
+
+keywords = ["quic", "http3", "webtransport", "media", "live"]
+categories = ["multimedia", "network-programming", "web-programming"]
+
+[dependencies]
+bytes = "1"
+thiserror = "2"
+tracing = "0.1"
+num_enum = "0.7"
