+++ moq-transfork-proto/src/coding/encode.rs
