+++ moq-transfork-proto/src/coding/varint.rs
