+++ moq-transfork-proto/src/message/fetch.rs
@@ -1,5 +1,5 @@
+use super::Path;
 use crate::coding::*;
-use crate::Path;
 
 #[derive(Clone, Debug)]
 pub struct Fetch {
