+++ moq-transfork-proto/src/message/mod.rs
@@ -7,6 +7,7 @@ mod fetch;
 mod frame;
 mod group;
 mod info;
+mod path;
 mod session;
 mod setup;
 mod stream;
@@ -19,6 +20,7 @@ pub use fetch::*;
 pub use frame::*;
 pub use group::*;
 pub use info::*;
+pub use path::*;
 pub use session::*;
 pub use setup::*;
 pub use stream::*;
