+++ moq-transfork/src/error.rs
@@ -1,4 +1,4 @@
-use crate::{coding, message};
+use moq_transfork_proto::message;
 
 /// A list of possible errors that can occur during the session.
 #[derive(thiserror::Error, Debug, Clone)]
@@ -7,65 +7,41 @@ pub enum Error {
 	WebTransport(#[from] web_transport::Error),
 
 	#[error("decode error: {0}")]
-	Decode(#[from] coding::DecodeError),
-
-	// TODO move to a ConnectError
-	#[error("unsupported versions: client={0:?} server={1:?}")]
-	Version(message::Versions, message::Versions),
-
-	/// A required extension was not present
-	#[error("extension required: {0}")]
-	RequiredExtension(u64),
-
-	/// An unexpected stream was received
-	#[error("unexpected stream: {0:?}")]
-	UnexpectedStream(message::ControlType),
+	Decode(#[from] moq_transfork_proto::coding::DecodeError),
 
 	/// Some VarInt was too large and we were too lazy to handle it
 	#[error("varint bounds exceeded")]
-	BoundsExceeded(#[from] coding::BoundsExceeded),
+	BoundsExceeded(#[from] moq_transfork_proto::coding::BoundsExceeded),
+
+	#[error("not found")]
+	NotFound,
+
+	// Cancel is returned when there are no more readers.
+	#[error("cancelled")]
+	Cancel,
 
 	/// A duplicate ID was used
 	// The broadcast/track is a duplicate
 	#[error("duplicate")]
 	Duplicate,
 
-	// Cancel is returned when there are no more readers.
-	#[error("cancelled")]
-	Cancel,
-
 	// The application closes the stream with a code.
 	#[error("app code={0}")]
 	App(u32),
-
-	#[error("not found")]
-	NotFound,
 
 	#[error("wrong frame size")]
 	WrongSize,
 
 	#[error("protocol violation")]
 	ProtocolViolation,
-}
 
-impl Error {
-	/// An integer code that is sent over the wire.
-	pub fn to_code(&self) -> u32 {
-		match self {
-			Self::Cancel => 0,
-			Self::RequiredExtension(_) => 1,
-			Self::WebTransport(_) => 4,
-			Self::Decode(_) => 5,
-			Self::Version(..) => 9,
-			Self::UnexpectedStream(_) => 10,
-			Self::BoundsExceeded(_) => 11,
-			Self::Duplicate => 12,
-			Self::NotFound => 13,
-			Self::WrongSize => 14,
-			Self::ProtocolViolation => 15,
-			Self::App(app) => *app + 64,
-		}
-	}
+	/// An unexpected stream was received
+	#[error("unexpected stream: {0:?}")]
+	UnexpectedStream(message::ControlType),
+
+	// TODO move to a ConnectError
+	#[error("unsupported versions: client={0:?} server={1:?}")]
+	Version(message::Versions, message::Versions),
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
