+++ moq-transfork/src/model/mod.rs
@@ -1,13 +1,11 @@
 mod announced;
 mod frame;
 mod group;
-mod path;
 mod router;
 mod track;
 
 pub use announced::*;
 pub use frame::*;
 pub use group::*;
-pub use path::*;
 pub use router::*;
 pub use track::*;
